Change the ISO upload host to IsoPomo, our new storage server.
In contrast to our previous storage server, the new one is not close to the builders and testers.
Uploading a bootcdregtest to it, just to download it again to all testers a minute later would not be efficient and slow down the entire building and testing process.
Therefore, bootcdregtest ISOs are now copied to and from a mount point in the buildfarm-local network. This is currently an NFSv4 mount, but can theoretically be anything.

The downside of this is that bootcdregtest ISOs are no longer available for download.
If that becomes a big issue, we can sync them to IsoPomo in a second step. But let's first figure out how important it actually is.