[WORKER] Reboot Windows testers after every run of `run_rostests`
This works by putting the new `buildbot_start.cmd` into Scheduled Tasks (run on logon).
It launches the BuildBot worker and waits for it to finish, then reboots the Windows VM.

The BuildBot worker is reconfigured to watch the `shutdown.stamp` file in the basedir, and ask the Buildmaster for a graceful shutdown of the worker if that file's timestamp has changed.
This works via BuildBot's allow_shutdown Worker config option.

All of this together should gracefully reboot the Windows VM after every test, even when multiple tests are queued and without causing any BuildBot exceptions (which always happen on unexpected shutdowns).

Dedicated to Timo, who asked me about this for months and also found the allow_shutdown option :)